Implement Player and PlayerStatus. Add relative requests and responses for WatchOut_Player.

+++ WatchOut_Player/src/main/java/beans/AddPlayerResponse.java
@@ -25,7 +25,7 @@ public class AddPlayerResponse
         this.playersEndpoints = playersEndpoints;
     }
 
-    public long getPlayerId()
+    public int getPlayerId()
     {
         return this.playerId;
     }
